tests/requirements-testing.txt
# =============================================================================
# CrackSeg Testing Dependencies
# =============================================================================
# Additional dependencies specifically for testing environment
# Used in Docker container for E2E and integration testing

# =============================================================================
# Testing Framework & Runners
# =============================================================================
pytest>=8.3.0
pytest-cov>=6.0.0
pytest-mock>=3.14.0
pytest-xdist>=3.6.0        # Parallel test execution
pytest-html>=4.1.1         # HTML test reports
pytest-metadata>=3.1.1     # Test metadata collection
pytest-timeout>=2.3.1      # Test timeout management
pytest-rerunfailures>=14.0 # Retry failed tests

# =============================================================================
# E2E Testing & Browser Automation
# =============================================================================
selenium>=4.28.0           # Web browser automation
webdriver-manager>=4.0.2   # Automatic webdriver management
requests>=2.32.0           # HTTP requests for API testing
beautifulsoup4>=4.12.3     # HTML parsing for content validation

# =============================================================================
# GUI Testing Specific
# =============================================================================
# streamlit-testing-lib>=0.1.0  # Streamlit testing utilities (if available) - NOT AVAILABLE
pynput>=1.7.7              # Keyboard/mouse automation (fallback)
pyautogui>=0.9.54          # Screen automation (backup testing)

# =============================================================================
# Screenshot & Visual Testing
# =============================================================================
Pillow>=10.4.0             # Image processing for screenshots
opencv-python>=4.11.0     # Computer vision for image comparison
numpy>=2.2.0               # Numerical operations for image analysis

# =============================================================================
# Test Data Management
# =============================================================================
faker>=33.1.0              # Generate test data
factory-boy>=3.3.1         # Test object factories
parameterized>=0.9.0       # Parameterized test cases

# =============================================================================
# Performance & Load Testing
# =============================================================================
locust>=2.32.4             # Load testing framework
memory-profiler>=0.61.0    # Memory usage profiling
psutil>=6.1.0              # System resource monitoring

# =============================================================================
# Test Configuration & Utilities
# =============================================================================
python-dotenv>=1.1.0       # Environment configuration
configparser>=7.1.0        # Configuration file handling
pyyaml>=6.0.2              # YAML configuration parsing

# =============================================================================
# Reporting & Documentation
# =============================================================================
allure-pytest>=2.13.5      # Advanced test reporting
coverage>=7.6.1            # Code coverage analysis
pytest-json-report>=1.5.0  # JSON test reports
jinja2>=3.1.6              # Template engine for reports

# =============================================================================
# Development & Debugging
# =============================================================================
pytest-xvfb>=3.0.0         # Virtual display for headless testing
pytest-env>=1.1.5          # Environment variable management
pytest-order>=1.3.0        # Test execution order control